experiment with Bitwise range analysis algorithm
The algorithm was used to minimize the number of bits needed for
FPGAs, described in the PLDI 2000 paper:

  Bitwidth Analysis with Applications to Silicon Compilation
  Mark Stephenson, Jonathan Babb, and Saman Amarasinghe

The algorithm is interesting for three reasons.

First, it extensively exploits undefined behavior.  For example,
given a[i], it concludes that i >= 0 and i < ARRAY_SIZE(a).  It
also concludes that signed a + b is at most INT_MAX (probably this
was due to saturating arithmetic rather than undefined behavior).

Second, it propagates value ranges in both directions, foward and
backward.  Backward propagation is crucial to optimize away the
zero check in the PostgreSQL division-by-zero case; neither GCC nor
Clang is able to do so (GCC only reorders the check).

Third, this is a general algorithm, which doesn't require specific
undefined behavior patterns.  It *just* works for all integer-based
examples in our APSys paper (two div-by-zero, oversized shift,
signed overflow, uninitialized read), better than GCC and Clang
combined.