insert llvm.debugtrap calls
Previously we inserted a call to int.sat() to check overflows for
expressions like x * y.

	int.sat(umul-overflow(x, y));

If the argument is optimized to 0, we avoid invoking the solver.

	int.sat(0);

The worry is that this special function int.sat() is not understood
by the LLVM optimizer, which may hinder further optimizations.  For
example, no instructions are allowed to be moved across this function
call since the optimizer considers it to have side effects.

We cannot mark int.sat() as readonly (pure) or readnone (const),
otherwise the optimizer removes it.

Now we change the rewriting to:

	if (umul-overflow(x, y))
		llvm.debugtrap();

And the checking pass simply evaluates if llvm.debugtrap() is
reachable.

Another option is to insert llvm.trap() instead, which is marked
as unreachable.  The difference is that this will add an extra
assumption that x * y doesn't overflow to future computations.