add -lower-logical
This pass breaks down composite logical conditions into branches,
so as to reduce false errors.  Consider the following example.

	x > UMAX / 4 && x * 4 > y

-simplifycfg will emits:

	%mul = %x, 4
	%cmp0 = icmp ugt %x, UMAX / 4
	%cmp1 = icmp ugt %mul, %y
	%cond = or %cmp0, %cmp1

This evaluates x * 4 without the precondition x > UMAX / 4, and
thus leads to a false overflow error.

A more tricky case comes with a swapped order in CVE-2011-1494.

	x * 4 > y && x > UMAX / 4

In this case, we need further to swap the order of the two comparisons
to avoid a false error.

This pass breaks AND/OR conditions into fine-brained branches.  A
-sink pass should run after this pass, to move down instructions
such as x * 4 in this example, so that further -simplifycfg cannot
revert the changes.