add signed range constraints
Since there's no sign information in the middle-end, we need to
carefully generate range constraints.

Given a ranged variable, we first assume it's signed, and compute
its signed min/max values.  Then we add constraints using signed
comparisons.

Again, we assume the variable is unsigned, and add unsigned range
constraints in a similar way.

A common case is that the range falls within the non-negative half,
such as [1, 100], a subset of [0, INT_MAX].  In this case it doesn't
matter if we use signed or unsigned comparisons.  A trivial
optimization is to generate signed constraints only.