anti-dce: fix FP in loops
Consider the following program with a loop.

	struct X *p;
	while (!(p = malloc(...))) {
		sleep(0);
	}
	p->a = 42;

There's no phi node generated in the code.

One could infer in this way:

1) p->a means p must be non-null;

2) The loop condition !p must be true.

3) sleep(0) is dead.

This is obviously wrong.  The subtle problem is that, though in the
def-use chain there is only a single p and p->a postdominates !p,
the two occurrences of p in and out of the loop are different: the
first p could be null while the second cannot.

We cannot use SSA in this way.  We need an extra instruction to
distinguish variables defined in a loop and used out of a loop.

One way to fix this is using loop-closed SSA (LCSSA), which inserts
an extra phi for p out of the loop.  Another (probably better) way
is to use gated SSA (GSA), which inserts η node at loop exit.

	p' = η(p, p != NULL)

However, inserting such nodes require to recognize loops, which is
tricky for programs with irreducible loops (e.g., via goto), because
LLVM only recognizes natural loops and doesn't work otherwise.

For now we use a more conservative approach.  If a statement is in
a loop, don't inspect its post-dominators at all.