anti-peephole: fix clone matching
This commit fixes two problems when doing pre-post diff.

1) The cloning scheme was wrong.  Only function arguments are
suppressed during cloning; we need to build the equivalance between
instructions.  For example, given a call to some function, we should
assume that the calls in two clones are equivalent.

2) a block's successors may be swapped when its branching conditions
is rewritten as the negation.  For example, p != null may be rewritten
as p == null, with two successors swapped.  We need do the mapping.

Not sure how to deal with rewriting of memory instructions.  A pass
may rewrite some memory loads; then it's hard to match memory loads
in two clones.

Any better idea than pre-post diff?  The -instcombine pass is still
too coarse-grained.  What we actually want is to simplify comparisons
only, not touching memory instructions at all.  Maybe we can try to
call SimplifyCmpInst().