add -simplify-delete
Clang emits a null pointer check "if (p)" for "delete p" in C++
code.  The check is considered "dead" if p was used earlier, which
leads to FPs.  This patch removes the debugging information on such
checks so that bugon calls will ignore them.

We could remove these checks completely (as gcc does) though.