reimplement anti-peephole
Given expression e, the basic idea is to ask an oracle for a
"simpler" form e' such that e' and e are only equivalent with
bug-free assertions.  Now the question is: what's the oracle?

This new implementation uses a "math" oracle, which is basically
an algebra system.  It represents both sides of a comparison,
such as (lhs < rhs), with symbolic polynomial expressions.  Then
the oracle transforms it into (lhs - rhs < 0) for "mathematical"
simplification (i.e., consider each variable as infinitely ranged).
For example, p + len < p then becomes len < 0.

The current problem is that it invokes LLVM's scev pass to deal
with symbolic expressions, and scev seems to have some nsw flag
missing problems on integer operations: it even cannot simplify
x + n < x to n < 0, where both n and x are signed integers, but
it works if x is a pointer since scev does that correctly.  In
other words, scev is too conservative for our purpose; we don't
really care about the oracle's correctness since we will verify
the equivalence later using SMT solvers anyway.

We could either fix scev or use another symbolic engine (e.g.,
GiNaC or Maple) as an oracle.