simplify-delete: use simpler heuristics
The problem is that Clang always emits a null check for "delete p"
in C++.  The previous appraoch is to look for calls to libcxx's
delete.  This doesn't work with overloaded delete.

This version uses a simpler rule: clang always emits such BBs with
the name "delete.notnull" which cannot be named by programmers.