anti-simplify: fix constant folding algorithm
The previous implementation is actually wrong, but happens to
work for many cases.  It works as follows.

1) Check if path condition R can be SAT (otherwise dead code).

2) Given an expression E, evaluate its value in the model when R
   is true, say C.

3) Check if E != C can be SAT (otherwise E is trivially constant).

4) With bug-free assertions Delta, check if E != C is UNSAT (i.e.,
   E must evaluate to C).

There are several problems.

First, E could have many values (in Step 2).  The above algorithm
only checks one of them and may miss optimizations in Step 4.  It
is better to obtain C with Delta.

Second, it is even unclear how to implement Step 2 correctly.  The
expression E is not part of the query (i.e., the path condition R),
so it is basically irrelevant in the model.  In that case, the SMT
solver responds with a do-not-care value in Step 2.

Consider p + 100 > p.  The solver may return any value, say 0, for
this comparison, when the path condition is trivially true.  Thus,
the algorithm fails to fold the expression to 1.

This implementation basically tries every possible constant value.
Fortunately, it only deals with boolean expressions for now (i.e.,
either true or false).