add signed shift test case
This is the runner up from the UB contest.  It also happened to
glibc's __ieee754_fmodf() function.  The corresponding rule is that
signed left shift x << y must have the same sign as that of x.

Note that currently clang doesn't emit the nsw flag for shl, and
thus we cannot really distinguish between signed and unsigned shifts.
This test uses sed to add the flag as a workaround.