reduce false positives due to assert(), BUG_ON(), etc
Crude -elim-assert pass to eliminate calls to assert-like functions,
in hopes of reducing the associated anti-simplify FPs.  It works, but
also creates some new FPs because we lose track of the condition from
the assert() call -- e.g., that some pointer is non-NULL.  So maybe
there is some other better way to do this?