add frontend plugin to kill macro FPs
This plugin removes debugging information of macro-expanded code,
such as BUG_ON(...), so that anti-* passes will not issue any
warnings for code from the macro.

To see the output of the plugin, use:

    clang -Xclang -load -Xclang lib/libintfe.so \
          -Xclang -plugin -Xclang intfe file.c \
          -c -g -o - | opt -S

This plugin works as follows:

1. Identify macro-expanded code by intercepting macro expansions in
   the preprocessor.

2. Use a wrapper to intercept the code generation action to get the
   generated module.

3. Modify the module using macro information.

Since we now ignore macro-expanded code, do we still need this
bugon-assert pass?