add command-line SMT backend
Instead of linking with the boolector library that is GPLed (and
with a more restricted license in the latest version), this backend
invokes an executable of an SMT solver.

	configure --with-smtlib=PATH-TO-YOUR-SOLVER

Here's an example of using the STP solver:

	configure --with-smtlib="/path/to/stp --SMTLIB2"

We may use the MIT license for our source code in this way.